Extract related images handling from search store
Related images store handles information about photos that are loaded below the image selected after search.
