+++ src/store/index.js
@@ -16,7 +16,7 @@ import UserStore from './user-store';
 import UsageDataStore from './usage-data-store';
 import FilterStore from './filter-store';
 import ReportContentStore from './report-content-store';
-
+import RelatedImagesStore from './related-images-store';
 Vue.use(Vuex);
 
 const queryParams = !(typeof window === 'undefined') ? window.location.search : '';
@@ -32,6 +32,7 @@ const store = (GoogleAnalytics, router) => (new Vuex.Store({
     SocialMediaStore.actions(GoogleAnalytics),
     ABTestStore.actions,
     ReportContentStore.actions(ReportService),
+    RelatedImagesStore.actions(ImageService),
   ),
   state: Object.assign(
     SearchStore.state(queryParams),
@@ -41,6 +42,7 @@ const store = (GoogleAnalytics, router) => (new Vuex.Store({
     ABTestStore.state,
     UserStore.state,
     ReportContentStore.state,
+    RelatedImagesStore.state,
   ),
   mutations: Object.assign(
     SearchStore.mutations(redirectTo(router)),
@@ -49,6 +51,7 @@ const store = (GoogleAnalytics, router) => (new Vuex.Store({
     BugReportStore.mutations,
     ABTestStore.mutations,
     ReportContentStore.mutations,
+    RelatedImagesStore.mutations,
   ),
 }));
